fix(discord): repair the test-helper find-replace that made the worker timeout unobservable
`createDiscordHandlerParams` in message-handler.test-helpers.ts carried a botched
find-replace — `// workerRunTimeoutMs: overrides?.// workerRunTimeoutMs,` — which
commented out the whole assignment and left a nested `//` mid-line. The effect was that
`workerRunTimeoutMs` was never threaded from the test harness into the handler params, so
the Discord inbound-worker timeout never fired and `runtime.error` was called 0 times no
matter what a test passed. The timeout was unreachable from tests.

Repair it to the intended form, mirroring the sibling lines. `DiscordMessageHandlerParams`
already declares `workerRunTimeoutMs?: number` and forwards it as the worker's
`runTimeoutMs`, so the source side needed no change.

The quarantine ledger's claim about this file was wrong and is corrected here. It said
message-handler.queue.test.ts stayed quarantined on exactly 3 timeout-fallback-reply cases;
it actually failed on 5. Two of those failed on the harness bug alone — "does not send the
timeout fallback when a final reply already went out" and "does not send the timeout
fallback when final reply delivery is already in flight" — both of which assert
`runtime.error` fires and the fallback is NOT sent. With the helper repaired they pass, and
the file drops from 5 failures to 3.

The 3 that remain are re-homed into message-handler.timeout-fallback-reply.test.ts
(the #2953/#2970 focused-spec precedent, as with the #2968 dedup re-home), so only that
focused file stays quarantined and message-handler.queue.test.ts un-gates and now blocks
merge on its 11 queue-behaviour cases.

`onTimeout` behaviour is deliberately untouched: it remains log-only. Whether the timeout
should deliver a user-facing reply is the separate, still-open decision in #2998 and is not
settled here.

Refs #2998
